Pull in weval branch of SpiderMonkey and add build.
This pulls in a new bytecodealliance/gecko-dev commit hash with weval
support, and produces a new separate build of the engine with intrinsics
added so that we can apply the [weval](https://github.com/cfallin/weval)
tool to obtain speedups with AOT compilation/specialization.

This is a re-do of #5, but with fixes from bytecodealliance/gecko-dev#5
applied.